refactor(backend): 抽出 SynthesisService，合成链路参数对象化；清理死代码并修复剩余审计问题
- SynthesisService 持有后台任务引用并在关停时取消；准备阶段失败不再让任务卡在 running；MP3 先写临时文件再改名
- tts：SynthesisParams 取代十余个逐层转发的参数，ChatTTS 依赖收进 handle，CPU 模式不再调用 cuda.synchronize，逐句 empty_cache 改为批级
- 合规审核：仅在输出不是合法 JSON 时重问，不再与 LLMClient 重试叠加；提示词与原始输出降为 DEBUG（含敏感对话）；LLM 重试覆盖 ConnectTimeout
- 运行中任务不允许 invalidate；/tasks 列表放入线程；OCR/YOLO 推理串行并加锁；健康检查改用公共访问器
- 删除无调用方的 correct/correct_segments/merge_chunks/group_segments/synthesize_dialogue 等及其测试，补 correct_transcript 测试；日志初始化改用 Settings
- download_models 同步缓存 MacBERT（离线生产环境需要）

+++ backend/src/copernicus/services/asr.py
@@ -642,7 +642,7 @@ class ASRService:
             return []
 
         try:
-            speech, sample_rate = sf.read(str(audio_path))
+            speech, sample_rate = sf.read(str(audio_path), dtype="float32")  # 默认 float64，长音频内存翻倍
         except Exception as e:
             logger.warning("Failed to read audio for diarization: %s", e)
             return []
